Fix loading of penstyle data
Summary: BUG:378561

Test Plan: Create report, insert a label/field/check/text, set linestyle, switch to data view

Reviewers: staniek

Tags: #kreport

Differential Revision: https://phabricator.kde.org/D5391

+++ src/common/KReportItemLine.cpp
@@ -53,9 +53,9 @@ KReportItemLine::KReportItemLine(const QDomNode & element)
         if (n == QLatin1String("report:line-style")) {
             KReportLineStyle ls;
             if (parseReportLineStyleData(node.toElement(), &ls)) {
-                m_lineWeight->setValue(ls.width());
+                m_lineWeight->setValue((int)ls.width());
                 m_lineColor->setValue(ls.color());
-                m_lineStyle->setValue(int(ls.penStyle()));
+                m_lineStyle->setValue(static_cast<int>(ls.penStyle()));
             }
         } else {
             kreportWarning() << "while parsing line element encountered unknown element: " << n;
